Move exhaustive tests for recovery to module
